+++ libraries/azure_network_security_group.rb
@@ -131,6 +131,9 @@ class AzureNetworkSecurityGroup < AzureGenericResource
     @security_rules ||= properties.securityRules
   end
 
+  def flow_log_retention_period
+    !properties.include?(:retentionPolicy) ? 0 : properties.retentionPolicy
+  end
   def default_security_rules
     return unless exists?
     @default_security_rules ||= properties.defaultSecurityRules
@@ -148,6 +151,20 @@ class AzureNetworkSecurityGroup < AzureGenericResource
   end
   RSpec::Matchers.alias_matcher :allow_rdp_from_internet, :be_allow_rdp_from_internet
 
+  def allow_udp_from_internet?
+    return unless exists?
+    allow_port_from_internet?('53')
+  end
+  RSpec::Matchers.alias_matcher :allow_udp_from_internet, :be_allow_udp_from_internet
+
+  SPECIFIC_CRITERIA = %i(specific_port access_allow direction_inbound source_open not_icmp).freeze
+
+  def allow_http_from_internet?
+    return unless exists?
+    allow_port_from_internet?('80') || allow_port_from_internet?('443')
+  end
+  RSpec::Matchers.alias_matcher :allow_http_from_internet, :be_allow_http_from_internet
+
   SPECIFIC_CRITERIA = %i(specific_port access_allow direction_inbound source_open not_icmp).freeze
   def allow_port_from_internet?(specific_port)
     return unless exists?
@@ -155,6 +172,8 @@ class AzureNetworkSecurityGroup < AzureGenericResource
     matches_criteria?(SPECIFIC_CRITERIA, security_rules_properties)
   end
   RSpec::Matchers.alias_matcher :allow_port_from_internet, :be_allow_port_from_internet
+
+
 
   private