+++ test/integration/verify/controls/azurerm_postgresql_server.rb
@@ -19,3 +19,16 @@ control 'azurerm_postgresql_server' do
     its('properties') { should have_attributes(version: '9.5') }
   end
 end
+
+control 'azure_postgresql_server' do
+
+  impact 1.0
+  title 'Checking the firewall rules for singular resource of azure_postgresql_server.'
+  desc 'Checking the firewall rules for singular resource of azure_postgresql_server.'
+
+  only_if { !postgresql_server_name.nil? }
+
+  describe azure_postgresql_server(resource_group: resource_group, server_name: postgresql_server_name) do
+    its('firewall_rules') { should eq {} }
+  end
+end